updating and adding mem functions

+++ strncpy.c
@@ -6,13 +6,14 @@
 /*   By: clopes <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/20 16:16:10 by clopes            #+#    #+#             */
-/*   Updated: 2019/05/22 11:50:09 by clopes           ###   ########.fr       */
+/*   Updated: 2019/06/03 11:35:31 by clopes           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
 char	*ftstrncpy(char *src, char *dst, size_t len)
 {
-	int k;
+size_t k;
 	k = 0;
 	while(dst[k] && k < len)
 	{
